chore: admin register component add
